pyproject.toml
[project]
name = "bbcoach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "altair<5",
    "beautifulsoup4>=4.14.3",
    "google-generativeai>=0.8.6",
    "pandas<3",
    "playwright>=1.58.0",
    "pyarrow>=23.0.0",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.41.0",
    "torch>=2.10.0",
    "transformers>=5.1.0",
]
